+++ common/socket-front/src/index.js
@@ -1,0 +1,6 @@
+import APILib from './API'
+import SocketLib from './Socket'
+
+export const API = APILib
+export const Socket = SocketLib
+