+++ spotify-analyzer/app/src/components/TitleInfosLine.vue
@@ -1,11 +1,11 @@
 <template>
   <Line
     :img="album?.images?.[0]?.url"
-    :infos="[
+    :infos="!onlyImg ? [
       ...additionalInfos,
       { icon: 'fas fa-user', text: artist?.name },
       { icon: 'fas fa-music', text: track?.name }
-    ]"
+    ] : []"
     />
 
 </template>
@@ -17,6 +17,7 @@ import Line from './Line.vue';
 
 const props = defineProps({
   trackId: { required: true, type: String },
+  onlyImg: { type: Boolean, default: false },
   additionalInfos: {
     /**@type {{icon:string, text: string}[]} */
     default: []
@@ -27,11 +28,11 @@ onMounted(async () => {
   await Dictionnary.addTrack(props.trackId)
 })
 
-/** @type {import('vue').Ref<import('@clabroche-org/spotify-analyzer-models/src/models/Track')>} */
+/** @type {import('vue').Ref<import('@clabroche/spotify-analyzer-models/src/models/Track')>} */
 const track = computed(() => Dictionnary.tracks.value[props.trackId])
-/** @type {import('vue').Ref<import('@clabroche-org/spotify-analyzer-models/src/models/Album')>} */
+/** @type {import('vue').Ref<import('@clabroche/spotify-analyzer-models/src/models/Album')>} */
 const album = computed(() => Dictionnary.albums.value[track.value?.albumId])
-/** @type {import('vue').Ref<import('@clabroche-org/spotify-analyzer-models/src/models/Artist')>} */
+/** @type {import('vue').Ref<import('@clabroche/spotify-analyzer-models/src/models/Artist')>} */
 const artist = computed(() => Dictionnary.artists.value[track.value?.artistsIds?.[0]])