+++ spotify-analyzer/app/src/views/Import.vue
@@ -1,0 +1,26 @@
+<template>
+  <div>
+    <input type="file" multiple @change="onchange">
+  </div>
+</template>
+
+<script setup>
+import { apis } from '@clabroche/spotify-analyzer-models'
+
+const onchange = (ev) => {
+  var formData = new FormData();
+  ;[...ev.target.files].forEach(file => {
+    console.log(file)
+    formData.append(file.name, file);
+  });
+  apis.coreAPI.instance.post('/api/histories/import', formData, {
+    headers: {
+      'Content-Type': 'multipart/form-data'
+    }
+  })
+}
+</script>
+
+<style lang="scss" scoped>
+
+</style>