+++ common/check-monorepo/CHANGELOG.md
@@ -3,9 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [1.1.1-beta.9](https://github.com/clabroche-org/monorepo/compare/v1.1.1-beta.8...v1.1.1-beta.9) (2022-08-05)
+## [1.1.1](https://github.com/clabroche/monorepo/compare/v1.1.1-beta.34...v1.1.1) (2022-12-04)
+
+**Note:** Version bump only for package @clabroche/common-check-monorepo
+
+## [1.1.1-beta.33](https://github.com/clabroche/monorepo/compare/v1.1.1-beta.32...v1.1.1-beta.33) (2022-12-04)
 
 ### Bug Fixes
 
-* **common-check-monorepo:** add script to check if monorepo is in good health ([a1c489c](https://github.com/clabroche-org/monorepo/commit/a1c489c25a9b4ec9a615e4a4c20e864dab5b8056))
-* **common-check-monorepo:** sync yarn install ([3ca7e7b](https://github.com/clabroche-org/monorepo/commit/3ca7e7b1615f022b9e49556d6a52e59a696b148c))
+* **owner:** change owner ([5fbde86](https://github.com/clabroche/monorepo/commit/5fbde863734787839cd409734ae5871dc5e1772f))
+
+## [1.1.1-beta.11](https://github.com/clabroche/monorepo/compare/v1.1.1-beta.10...v1.1.1-beta.11) (2022-10-16)
+
+### Bug Fixes
+
+* **spotify-analyzer:** add it ([4566276](https://github.com/clabroche/monorepo/commit/45662766b3fe3e12930ecb73d97d27424eb701b2))
+
+## [1.1.1-beta.9](https://github.com/clabroche/monorepo/compare/v1.1.1-beta.8...v1.1.1-beta.9) (2022-08-05)
+
+### Bug Fixes
+
+* **common-check-monorepo:** add script to check if monorepo is in good health ([a1c489c](https://github.com/clabroche/monorepo/commit/a1c489c25a9b4ec9a615e4a4c20e864dab5b8056))
+* **common-check-monorepo:** sync yarn install ([3ca7e7b](https://github.com/clabroche/monorepo/commit/3ca7e7b1615f022b9e49556d6a52e59a696b148c))