+++ common/express/CHANGELOG.md
@@ -3,12 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [1.1.1-beta.3](https://github.com/clabroche-org/monorepo/compare/v1.1.1-beta.2...v1.1.1-beta.3) (2022-08-03)
+## [1.1.1](https://github.com/clabroche/monorepo/compare/v1.1.1-beta.34...v1.1.1) (2022-12-04)
 
-**Note:** Version bump only for package @clabroche-org/common-express
+**Note:** Version bump only for package @clabroche/common-express
+
+## [1.1.1-beta.33](https://github.com/clabroche/monorepo/compare/v1.1.1-beta.32...v1.1.1-beta.33) (2022-12-04)
+
+### Bug Fixes
+
+* **owner:** change owner ([5fbde86](https://github.com/clabroche/monorepo/commit/5fbde863734787839cd409734ae5871dc5e1772f))
+
+## [1.1.1-beta.3](https://github.com/clabroche/monorepo/compare/v1.1.1-beta.2...v1.1.1-beta.3) (2022-08-03)
+
+**Note:** Version bump only for package @clabroche/common-express
 
 ## 1.1.1-beta.0 (2022-07-22)
 
 ### Bug Fixes
 
-* reset ([f6f7c95](https://github.com/clabroche-org/monorepo/commit/f6f7c951528ccd69d917a951f496ef8a308183c8))
+* reset ([f6f7c95](https://github.com/clabroche/monorepo/commit/f6f7c951528ccd69d917a951f496ef8a308183c8))